# src/api.js
const express = require("express");
const router = express.Router();

// Endpoint para criar um token
router.post("/token/create", (req, res) => {
  const { value } = req.body;
  if (!value) {
    return res.status(400).json({ error: "Valor do token é necessário" });
  }
  res.status(200).json({ message: "Token criado com sucesso", value });
});

// Endpoint para transferir tokens
router.post("/token/transfer", (req, res) => {
  const { from, to, amount } = req.body;
  if (!from || !to || !amount) {
    return res
      .status(400)
      .json({ error: "Informações de transferência são necessárias" });
  }
  res
    .status(200)
    .json({ message: "Transferência realizada com sucesso", from, to, amount });
});

// Endpoint para calcular o valor do token
router.post("/token/predict-value", (req, res) => {
  const { bureauData } = req.body;
  if (!bureauData) {
    return res
      .status(400)
      .json({ error: "Dados do bureau são necessários para predição" });
  }

  const predictedValue = calculateTokenValue(bureauData);
  res.status(200).json({ predictedValue });
});

// Função mockada para cálculo do valor do token
const calculateTokenValue = (bureauData) => {
  // Lógica real de predição seria implementada aqui
  return 100; // Valor de exemplo
};

module.exports = router;